Factor out form-context.tsx.
